Refactor call control rendering
